requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
bcrypt==4.0.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2024.2.0
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
huggingface-hub==0.20.3
idna==3.4
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
motor==3.3.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
passlib==1.7.4
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tokenizers==0.15.1
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.8.0
urllib3==2.2.0
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
